consolidate/address feedback
